blog/djangounchained/requirements.txt
amqp==2.3.2
anyjson==0.3.3
billiard==3.5.0.5
celery==4.2.1
Django==2.1.2
django-extensions==2.1.2
django-widget-tweaks==1.4.3
kombu==4.2.2.post1
Pillow==5.3.0
psycopg2==2.7.5
psycopg2-binary==2.7.5
python-dotenv==0.9.1
pytz==2018.5
six==1.11.0
vine==1.1.4
